refac(rs/base): rename `rust_vec_copyable.h` to `rust_vec.h`
This makes more sense because this source file also contains the definition
of `RustVec` and `Copyable` is just a part of it.
