+++ tachyon/rs/base/BUILD.bazel
@@ -14,8 +14,8 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "rust_vec_copyable",
-    hdrs = ["rust_vec_copyable.h"],
+    name = "rust_vec",
+    hdrs = ["rust_vec.h"],
     deps = [
         "//tachyon:export",
         "//tachyon/base:logging",
